Simplify

+++ spectest/bulk/table_fill/table_fill.0.go
@@ -25,7 +25,7 @@ func (m *Module) Xfill_abbrev(v0 int32, v1 any, v2 int32) {
 }
 func (m *Module) Xget(v0 int32) any {
 	t0 := v0
-	t1 := m.t0[uint(t0)]
+	t1 := m.t0[t0]
 	return t1
 }
